Make Array be U16 based.

+++ server/prototypes/rust/array/src/integration.c
@@ -9,13 +9,13 @@ rust_array_new(void);
 extern void
 rust_array_free(rust_array_t *);
 
-extern uint32_t
-rust_array_append(const rust_array_t *, uint32_t item);
+extern uint16_t
+rust_array_append(const rust_array_t *, uint16_t item);
 
-extern uint32_t
+extern uint16_t
 rust_array_first(const rust_array_t *);
 
-extern uint32_t
+extern uint16_t
 rust_array_last(const rust_array_t *);
 
 int main(void) {}
